Finished the refactor, tested for cv model, need to test for nlp
