+++ tests/sparsezoo/refactor/__init__.py
@@ -1,5 +1,3 @@
-# flake8: noqa
-
 # Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,27 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-- Functionality for accessing models, recipes, and supporting files in the SparseZoo
-- Notify the user the last pypi package version
-"""
-
-# flake8: noqa
-# isort: skip_file
-
-from .version import *
-from .main import *
-from .download_main import *
-from .models.zoo import *
-from .objects import *
-from .package import *
-from .refactor import *
-
-
-from sparsezoo.package import check_package_version as _check_package_version
-
-_check_package_version(
-    package_name=__name__ if is_release else f"{__name__}-nightly",
-    package_version=version,
-)
